refactor: command service는 command 패키지로 분리
